move 'write_toolchain_files' to base class
